SHRINKWRAP-183 - Rework configuration mechanism for type mappings to default extensions
Moved the ExtensionMapping map from Configuration to ExtensionLoader.
Added method to ExtensionLoader to get the extension from the mapping.
Added ExtensionWrapper, and changed the extensionMapping to contain it as the value field.
Updated all META-INF/services files. 
The loading of services files now throws a pretty exception if the file does not contain implementingClassName property.
The ServiceExtensionLoader now exposes the extension via getExtensionFromExtensionMapping. It also initializes the extensionMapping.
Added a delegator to be able to use UnknownExtensionTypeException from impl-base. Also removed code in UnknownExtensionTypeException that was no longer relevant.