SHRINKWRAP-505 - Allow addPackage() to be smarter about .war files
This commit brings about the smartness required for URLPackageScanner
to understand that a URL that represents a .war has its classes
relocated under WEB-INF/lib instead of from the root of the archive.

Part of the change is pre-constructing the Asset passed to the
scanner's Callback, taking into account the location of the underlying
asset, separate from it's actual related class name.